feat: can set timeout when requesting utxo server (#163)
One can now set:
* `DialTimeout` that will make the request fail if opening the
  connection with the server takes too long;
* `RequestTimeout` if the server is taking too much time to
  answer requests.

Before that, the only way to set these timeout was to fully
override the `HTTPClient` with a custom one, loosing the
value from `createHttpClient`. One can still provide their
own `HTTPClient` but if the need is only to customize these
timeouts, this change makes it easier.